make it noir module
