+++ packages/compiler/src/noir.rs
@@ -1,0 +1,100 @@
+use std::{fs::File, io::Write, path::Path};
+
+use crate::structs::RegexAndDFA;
+
+pub fn gen_noir_fn(regex_and_dfa: &RegexAndDFA, path: &Path) -> Result<(), std::io::Error> {
+    let noir_fn = to_noir_fn(regex_and_dfa);
+    let mut file = File::create(path)?;
+    file.write_all(noir_fn.as_bytes())?;
+    file.flush()?;
+    Ok(())
+}
+
+fn to_noir_fn(regex_and_dfa: &RegexAndDFA) -> String {
+    let accept_state_id = {
+        let last_state = regex_and_dfa.dfa.states.last().expect("no last state");
+        assert!(
+            last_state.state_type == "accept",
+            "last state is accept, right??"
+        );
+        last_state.state_id
+    };
+
+    const BYTE_SIZE: u32 = 256; // u8 size
+    let mut lookup_table_body = String::new();
+
+    // curr_state + char_code -> next_state
+    let mut rows: Vec<(usize, u8, usize)> = vec![];
+
+    for state in regex_and_dfa.dfa.states.iter() {
+        if state.state_type == "accept" {
+            assert_eq!(state.transitions.len(), 0, "accept state has transitions");
+        } else {
+            assert!(state.transitions.len() > 0, "no transitions");
+            for (&tran_next_state_id, tran) in &state.transitions {
+                for &char_code in tran {
+                    rows.push((state.state_id, char_code, tran_next_state_id));
+                }
+            }
+        };
+    }
+
+    for (curr_state_id, char_code, next_state_id) in rows {
+        lookup_table_body +=
+            &format!("table[{curr_state_id} * {BYTE_SIZE} + {char_code}] = {next_state_id};\n",);
+    }
+
+    lookup_table_body = indent(&lookup_table_body);
+    let table_size = BYTE_SIZE as usize * regex_and_dfa.dfa.states.len();
+    let lookup_table = format!(
+        r#"
+comptime fn make_lookup_table() -> [Field; {table_size}] {{
+    let mut table = [0; {table_size}];
+{lookup_table_body}
+
+    // experimentally confirmed that storing a transition for each char code for accept state produces less gates than adding an `if` to check if the current state is not "accept"
+    // I might be wrong. I tested for input of length 128 and 1024.
+    for i in 0..{BYTE_SIZE} {{
+        table[{accept_state_id} * {BYTE_SIZE} + i] = {accept_state_id};
+    }}
+    table
+}}
+    "#
+    );
+
+    let fn_body = format!(
+        r#"
+global table = make_lookup_table();
+pub fn regex_match<let N: u32>(input: [u8; N]) {{
+    // regex: {regex_pattern}
+    let mut s = 0;
+    for i in 0..input.len() {{
+        s = table[s * {BYTE_SIZE} + input[i] as Field];
+    }}
+    assert_eq(s, {accept_state_id}, f"no match: {{s}}");
+}}
+    "#,
+        regex_pattern = regex_and_dfa.regex_pattern,
+    );
+    format!(
+        r#"
+        {fn_body}
+        {lookup_table}
+    "#
+    )
+    .trim()
+    .to_owned()
+}
+
+fn indent(s: &str) -> String {
+    s.split("\n")
+        .map(|s| {
+            if s.trim().is_empty() {
+                s.to_owned()
+            } else {
+                format!("{}{}", "    ", s)
+            }
+        })
+        .collect::<Vec<_>>()
+        .join("\n")
+}
